refactor: rename section-folder
